+++ landingzones/tests/test_permissions.py
@@ -9,8 +9,6 @@ from projectroles.tests.test_permissions import (
     TestPermissionMixin,
 )
 
-from landingzones.tests.test_views_taskflow import LandingZoneTaskflowMixin
-
 # Samplesheets dependency
 from samplesheets.tests.test_io import SampleSheetIOMixin, SHEET_DIR
 
@@ -38,7 +36,6 @@ TEST_OBJ_NAME = 'test1.txt'
 
 class TestLandingZonePermissionsBase(
     LandingZoneMixin,
-    LandingZoneTaskflowMixin,
     SampleSheetIOMixin,
     TestPermissionMixin,
     TaskflowAPIPermissionTestBase,
@@ -61,15 +58,6 @@ class TestLandingZonePermissionsBase(
             configuration=None,
             config_data={},
         )
-        # Create zone in taskflow
-        self.make_zone_taskflow(self.landing_zone)
-        # Get collections
-        self.zone_coll = self.irods.collections.get(
-            self.irods_backend.get_path(self.landing_zone)
-        )
-        # Add files to zone
-        self.irods_obj = self.make_object(self.zone_coll, TEST_OBJ_NAME)
-        self.make_md5_object(self.irods_obj)
 
 
 class TestLandingZonePermissions(TestLandingZonePermissionsBase):
@@ -201,316 +189,6 @@ class TestLandingZonePermissions(TestLandingZonePermissionsBase):
         """Test ZoneCreateView with disabled non-superuser access"""
         url = reverse(
             'landingzones:create', kwargs={'project': self.project.sodar_uuid}
-        )
-        good_users = [self.superuser]
-        bad_users = [
-            self.user_owner_cat,
-            self.user_delegate_cat,
-            self.user_contributor_cat,
-            self.user_guest_cat,
-            self.user_finder_cat,
-            self.user_owner,
-            self.user_delegate,
-            self.user_contributor,
-            self.user_guest,
-            self.user_no_roles,
-            self.anonymous,
-        ]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
-
-    def test_zone_update(self):
-        """Test ZoneUpdateView permissions"""
-        url = reverse(
-            'landingzones:update',
-            kwargs={'landingzone': self.landing_zone.sodar_uuid},
-        )
-        good_users = [
-            self.superuser,
-            self.user_owner_cat,
-            self.user_delegate_cat,
-            self.user_owner,
-            self.user_delegate,
-            self.user_contributor,
-        ]
-        bad_users = [
-            self.user_contributor_cat,
-            self.user_guest_cat,
-            self.user_finder_cat,
-            self.user_guest,
-            self.user_no_roles,
-            self.anonymous,
-        ]
-        self.assert_response(url, good_users, 200)
-        redirect_url = reverse(
-            'landingzones:list',
-            kwargs={'project': self.landing_zone.project.sodar_uuid},
-        )
-        self.assert_response(url, bad_users, 302, redirect_user=redirect_url)
-
-    def test_zone_update_archive(self):
-        """Test ZoneUpdateView with archived project"""
-        self.project.set_archive()
-        url = reverse(
-            'landingzones:update',
-            kwargs={'landingzone': self.landing_zone.sodar_uuid},
-        )
-        good_users = [
-            self.superuser,
-        ]
-        bad_users = [
-            self.user_owner_cat,
-            self.user_delegate_cat,
-            self.user_contributor_cat,
-            self.user_guest_cat,
-            self.user_finder_cat,
-            self.user_owner,
-            self.user_delegate,
-            self.user_contributor,
-            self.user_guest,
-            self.user_no_roles,
-            self.anonymous,
-        ]
-        self.assert_response(url, good_users, 200)
-        redirect_url = reverse(
-            'landingzones:list',
-            kwargs={'project': self.landing_zone.project.sodar_uuid},
-        )
-        self.assert_response(url, bad_users, 302, redirect_user=redirect_url)
-
-    @override_settings(LANDINGZONES_DISABLE_FOR_USERS=True)
-    def test_zone_update_disable(self):
-        """Test ZoneUpdateView with disabled non-superuser access"""
-        url = reverse(
-            'landingzones:update',
-            kwargs={'landingzone': self.landing_zone.sodar_uuid},
-        )
-        good_users = [
-            self.superuser,
-        ]
-        bad_users = [
-            self.user_owner_cat,
-            self.user_delegate_cat,
-            self.user_contributor_cat,
-            self.user_guest_cat,
-            self.user_finder_cat,
-            self.user_owner,
-            self.user_delegate,
-            self.user_contributor,
-            self.user_guest,
-            self.user_no_roles,
-            self.anonymous,
-        ]
-        self.assert_response(url, good_users, 200)
-        redirect_url = reverse(
-            'landingzones:list',
-            kwargs={'project': self.landing_zone.project.sodar_uuid},
-        )
-        self.assert_response(url, bad_users, 302, redirect_user=redirect_url)
-
-    def test_zone_delete(self):
-        """Test ZoneDeleteView permissions"""
-        url = reverse(
-            'landingzones:delete',
-            kwargs={'landingzone': self.landing_zone.sodar_uuid},
-        )
-        good_users = [
-            self.superuser,
-            self.user_owner_cat,
-            self.user_delegate_cat,
-            self.user_owner,
-            self.user_delegate,
-            self.user_contributor,
-        ]
-        bad_users = [
-            self.user_contributor_cat,
-            self.user_guest_cat,
-            self.user_finder_cat,
-            self.user_guest,
-            self.user_no_roles,
-            self.anonymous,
-        ]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
-
-    def test_zone_delete_archive(self):
-        """Test ZoneDeleteView with archived project"""
-        self.project.set_archive()
-        url = reverse(
-            'landingzones:delete',
-            kwargs={'landingzone': self.landing_zone.sodar_uuid},
-        )
-        good_users = [
-            self.superuser,
-            self.user_owner_cat,
-            self.user_delegate_cat,
-            self.user_owner,
-            self.user_delegate,
-            self.user_contributor,
-        ]
-        bad_users = [
-            self.user_contributor_cat,
-            self.user_guest_cat,
-            self.user_finder_cat,
-            self.user_guest,
-            self.user_no_roles,
-            self.anonymous,
-        ]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
-
-    @override_settings(LANDINGZONES_DISABLE_FOR_USERS=True)
-    def test_zone_delete_disable(self):
-        """Test ZoneDeleteView with disabled non-superuser access"""
-        url = reverse(
-            'landingzones:delete',
-            kwargs={'landingzone': self.landing_zone.sodar_uuid},
-        )
-        good_users = [self.superuser]
-        bad_users = [
-            self.user_owner_cat,
-            self.user_delegate_cat,
-            self.user_contributor_cat,
-            self.user_guest_cat,
-            self.user_finder_cat,
-            self.user_owner,
-            self.user_delegate,
-            self.user_contributor,
-            self.user_guest,
-            self.user_no_roles,
-            self.anonymous,
-        ]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
-
-    def test_zone_move(self):
-        """Test ZoneMoveView permissions"""
-        url = reverse(
-            'landingzones:move',
-            kwargs={'landingzone': self.landing_zone.sodar_uuid},
-        )
-        good_users = [
-            self.superuser,
-            self.user_owner_cat,
-            self.user_delegate_cat,
-            self.user_owner,
-            self.user_delegate,
-            self.user_contributor,
-        ]
-        bad_users = [
-            self.user_contributor_cat,
-            self.user_guest_cat,
-            self.user_finder_cat,
-            self.user_guest,
-            self.user_no_roles,
-            self.anonymous,
-        ]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
-
-    def test_zone_move_archive(self):
-        """Test ZoneMoveView with archived project"""
-        self.project.set_archive()
-        url = reverse(
-            'landingzones:move',
-            kwargs={'landingzone': self.landing_zone.sodar_uuid},
-        )
-        good_users = [self.superuser]
-        bad_users = [
-            self.user_owner_cat,
-            self.user_delegate_cat,
-            self.user_contributor_cat,
-            self.user_guest_cat,
-            self.user_finder_cat,
-            self.user_owner,
-            self.user_delegate,
-            self.user_contributor,
-            self.user_guest,
-            self.user_no_roles,
-            self.anonymous,
-        ]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
-
-    @override_settings(LANDINGZONES_DISABLE_FOR_USERS=True)
-    def test_zone_move_disable(self):
-        """Test ZoneMoveView with disabled non-superuser access"""
-        url = reverse(
-            'landingzones:move',
-            kwargs={'landingzone': self.landing_zone.sodar_uuid},
-        )
-        good_users = [self.superuser]
-        bad_users = [
-            self.user_owner_cat,
-            self.user_delegate_cat,
-            self.user_contributor_cat,
-            self.user_guest_cat,
-            self.user_finder_cat,
-            self.user_owner,
-            self.user_delegate,
-            self.user_contributor,
-            self.user_guest,
-            self.user_no_roles,
-            self.anonymous,
-        ]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
-
-    def test_zone_validate(self):
-        """Test ZoneMoveView for zone validation"""
-        url = reverse(
-            'landingzones:validate',
-            kwargs={'landingzone': self.landing_zone.sodar_uuid},
-        )
-        good_users = [
-            self.superuser,
-            self.user_owner_cat,
-            self.user_delegate_cat,
-            self.user_owner,
-            self.user_delegate,
-            self.user_contributor,
-        ]
-        bad_users = [
-            self.user_contributor_cat,
-            self.user_guest_cat,
-            self.user_finder_cat,
-            self.user_guest,
-            self.anonymous,
-            self.user_no_roles,
-        ]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
-
-    def test_zone_validate_archive(self):
-        """Test ZoneMoveView for zone validation with archived project"""
-        self.project.set_archive()
-        url = reverse(
-            'landingzones:validate',
-            kwargs={'landingzone': self.landing_zone.sodar_uuid},
-        )
-        good_users = [self.superuser]
-        bad_users = [
-            self.user_owner_cat,
-            self.user_delegate_cat,
-            self.user_contributor_cat,
-            self.user_guest_cat,
-            self.user_finder_cat,
-            self.user_owner,
-            self.user_delegate,
-            self.user_contributor,
-            self.user_guest,
-            self.user_no_roles,
-            self.anonymous,
-        ]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
-
-    @override_settings(LANDINGZONES_DISABLE_FOR_USERS=True)
-    def test_zone_validate_disable(self):
-        """Test ZoneMoveView for zone validation with disabled non-superuser access"""
-        url = reverse(
-            'landingzones:validate',
-            kwargs={'landingzone': self.landing_zone.sodar_uuid},
         )
         good_users = [self.superuser]
         bad_users = [
